+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- MSRV is now 1.63.0.
+- Migrated to `ff 0.14`, `rand_core 0.9`.
+- `group::Group::random(rng: impl RngCore) -> Self` has been changed to
+  `Group::random<R: RngCore + ?Sized>(rng: &mut R) -> Self`, to enable passing a
+  trait object as the RNG.
+- `group::Group::try_from_rng` is a new trait method that must be implemented by
+  downstreams. `Group::random` now has a default implementation that calls it.
 
 ## [0.13.0] - 2022-12-06
 ### Changed